Initial vite setup
