+++ src/main.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+import ITLWidget from "./components/ITLWidget";
+
+ReactDOM.createRoot(document.getElementById("root")).render(
+  <React.StrictMode>
+    <ITLWidget />
+  </React.StrictMode>
+);
