+++ src/widget.js
