--- widget/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8" />
-    <link hrel="stylesheet" ref="https://fonts.cdnfonts.com/css/bahnschrift">
-    <link rel="stylesheet" href="themes/dark/style.css">
-    <title>International Timing League</title>
-  </head>
-  <body>
-    <div class="bgwrap">
-      <div class="entrant"></div>
-    </div>
-    <script src="https://unpkg.com/react@16/umd/react.production.min.js" crossorigin></script>
-    <script src="https://unpkg.com/react-dom@16/umd/react-dom.production.min.js" crossorigin></script>
-    <script src="widget.js"></script>
-  </body>
-</html>
